requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.2

# For transcription
google-generativeai>=0.3.1

# For RAG recommendations
chromadb>=0.4.18
groq>=0.4.0
sentence-transformers>=2.2.2

# For Zoom API integration
requests>=2.28.0
python-dotenv>=0.19.0

# Utilities
aiofiles>=0.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# For speaker diarization
pyannote.audio>=2.1.1
torch>=2.0.0
torchaudio>=2.0.0

# UI
# tkinter is part of standard library and doesn't need to be installed 